Fixed a bug causing some video files to fail uploading in the Load Video UI node.
